+++ CHANGELOG.md
@@ -111,7 +111,8 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Renamed item stack network id to unique id for better readability, changes are made to multiple related methods.
 - (API) Refactored the recipe system, network-related code is moved to the server module.
 - (API) Moved package `item.descriptor` to `item.recipe.descriptor` since the item descriptor is only used in recipe system.
-- (API) Replaced `ModalFormCancelReason` with `FormCancelReason` in the modal form APIs.
+- (API) Replaced protocol library class `ModalFormCancelReason` with `FormCancelReason` in the modal form APIs.
+- (API) Replaced protocol library class `CreativeItemCategory` with `CreativeItemCategory.Type` in the creative item APIs.
 
 ### Fixed